attempt gym->gymnasium
This is as far as I got, but it's broken for some reason - the game
doesn't actually run. Feel free to revert if necessary.

Changes:

env.reset() must accept an optional 'seed' kwarg, and return a 2nd value
'info':
https://gymnasium.farama.org/main/introduction/migration_guide/#environment-reset

env.step() must now return 'terminated' and 'truncated' in place of
'done':
https://gymnasium.farama.org/main/introduction/migration_guide/#environment-step